[cli] separate generate-abi command
