+++ cli/src/cmd/mod.rs
@@ -14,12 +14,22 @@
 // You should have received a copy of the GNU General Public License
 // along with ink!.  If not, see <http://www.gnu.org/licenses/>.
 
+use std::{
+    io::{
+        self,
+        Write,
+    },
+    process::Command,
+};
+
+mod abi;
 mod build;
 mod deploy;
 mod error;
 mod new;
 
 pub(crate) use self::{
+    abi::execute_generate_abi,
     build::execute_build,
     deploy::execute_deploy,
     error::{
@@ -28,3 +38,20 @@ pub(crate) use self::{
     },
     new::execute_new,
 };
+
+fn exec_cargo(command: &str, args: &[&'static str]) -> Result<()> {
+    let output = Command::new("cargo")
+        .arg("+nightly")
+        .arg(command)
+        .args(args)
+        .output()?;
+
+    if !output.status.success() {
+        // Dump the output streams produced by cargo into the stdout/stderr.
+        io::stdout().write_all(&output.stdout)?;
+        io::stderr().write_all(&output.stderr)?;
+        return Err(error::CommandError::BuildFailed)
+    }
+
+    Ok(())
+}
